# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"RAG Agent API","description":"API for RAG agent integration with frontend","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Read Root","description":"Root endpoint to verify API is running","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/chat/completions":{"post":{"summary":"Chat Completions","description":"Chat completions endpoint that uses the RAG agent to generate responses\nbased on textbook content.","operationId":"chat_completions_api_v1_chat_completions_post","parameters":[{"name":"request","in":"query","required":true,"schema":{"title":"Request"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"selected_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Selected Text"},"session_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Token"},"mode":{"type":"string","title":"Mode","default":"full_content"}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response"},"session_token":{"type":"string","title":"Session Token"},"query":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Query"},"context_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context Used"}},"type":"object","required":["response","session_token"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}